pyproject.toml
[project]
name = "cyoa-tui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.5",
    "huggingface-hub[cli]==0.24.6",
    "jinja2>=3.1.6",
    "jiter>=0.13.0",
    "llama-cpp-python>=0.3.16",
    "neo4j>=6.1.0",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "textual>=8.0.2",
    "tiktoken>=0.12.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.7",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "no_worker_mock: skip the autouse mock_textual_workers fixture for tests using app.run_test()",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by ruff format)
]

[tool.ruff.lint.isort]
known-first-party = ["cyoa"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
